More doc, comments and remove specific effects instead of all
Now removes the specific potion effects applied by the set instead of removing everything. This way the player keeps their other potion effects not related to the set.
Change from ArmorItem to Item so that any item is accepted, in anticipation if we ever want to have main and off-hand items to be counted in the armor as well or not
Added a lot of comments to explain what is going on