Added ingame command to reload json file
Added command `/realisticarmortiers reload` to read from equipment_sets.json the armor sets and store them in our parser. This allows for users/modders to change the json file while in-game and just run the command to reload the file, preventing them from having to quit and restart the client to see the changes.
Factored some code in the RealisticArmorTiers, with more comments/explanations, these are still testing and will probably be removed for another completely different approach/implementation (I don't see how we can go further from here with this approach)